+++ minigist/pure_client.py
@@ -81,7 +81,7 @@ class PureMDClient:
         )
         return base_url_normalized + target_url
 
-    def fetch_markdown_content(self, target_url: str, timeout: int = DEFAULT_HTTP_TIMEOUT_SECONDS) -> str | None:
+    def fetch_markdown_content(self, target_url: str, timeout: float = DEFAULT_HTTP_TIMEOUT_SECONDS) -> str | None:
         if not self.api_token:
             self._apply_rate_limit_delay_if_needed()
